+++ src/agent/implementations/flows/tooluse/nodes/index.ts
@@ -1,0 +1,18 @@
+/**
+ * Tool-use flow nodes.
+ *
+ * Exports all node implementations for the tool-use flow.
+ */
+export { ToolUsePrepareNode } from './ToolUsePrepareNode';
+export { ToolUseCycleNode } from './ToolUseCycleNode';
+export { ToolUseWaitNode } from './ToolUseWaitNode';
+export type {
+  ToolUseRunShared,
+  StateSlicesSnapshot,
+  PrepareResult,
+  CycleExecResult,
+  WaitExecResult,
+  CyclePrepResult,
+  PreparedShared,
+} from './types';
+export { assertPreparedShared } from './types';
